Moved game UI into a package. Fixed some warnings.
