+++ src/main/java/gui/game/TwoTruthsAndALiePagePanelInterface.java
@@ -1,10 +1,10 @@
-package gui;
+package gui.game;
 
 import javax.swing.*;
 import java.io.IOException;
 
 public interface TwoTruthsAndALiePagePanelInterface {
-    public void update() throws IOException;
+    void update() throws IOException;
 
     void addGame(String username) throws IOException;
