+++ src/main/java/gui/game/TwoTruthsAndALiePageScrollPane.java
@@ -1,4 +1,6 @@
-package gui;
+package gui.game;
+import gui.MainFrame;
+
 import javax.swing.*;
 import java.awt.*;
